requirements.txt
awscli==1.18.56
boto3==1.13.7
botocore==1.16.6
certifi==2020.4.5.1
docutils==0.15.2
jmespath==0.9.5
pyasn1==0.4.8
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3.1
PyMySQL==0.9.3
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.9
